further simplified code and added styling
reworked more of the functions within the js file. added more styling to better see what I'm doing in the html. changed html input from number to text so that the number chosen sat more squarely in the input pane (also removed form that caused enter-key to submit the number in the input)