added another html file for final product
added another html file to serve as the basis for the final product. index.html was not shaping up the way that i wanted. decided to scrap and start over a new index file to serve as a foundation for how i want the site to look and behave eventually