new index.HTML file and updated JS
Added a 2nd HTML file to try starting over from my previous attempt at a layout. This time using Bootstrap.
Updated JS file to have dice roller function look to see if any of the inputs are either blank or NaN so that the output doesn't throw an error. Wrote lines for beginnings of pushing dice Outcome into Cards in the HTML